Refactor client/WSDL usage and resource access
