+++ MANIFEST.in
@@ -1,3 +1,2 @@
 include *.txt
-include correios/wsdls/*
 graft correios/data
