+++ correios/update_wsdl.py
@@ -1,88 +1,71 @@
+import argparse
 import logging
-import os
+import sys
 
+import os
 import requests
 
 logger = logging.getLogger(__name__)
 
-
-MODULE_PATH = os.path.join(os.path.dirname(__file__), 'wsdls')
-
-
-SPECS = [
-    {
-        'url': (
-            'https://apps.correios.com.br/'
-            'SigepMasterJPA/AtendeClienteService/AtendeCliente?wsdl'
-        ),
-        'filename': 'AtendeCliente-production.wsdl'
-    },
-    {
-        'url': (
-            'https://apphom.correios.com.br/'
-            'SigepMasterJPA/AtendeClienteService/AtendeCliente?wsdl'
-        ),
-        'filename': 'AtendeCliente-test.wsdl'
-    },
-    {
-        'url': (
-            'https://webservice.correios.com.br/'
-            'service/rastro/Rastro.wsdl'
-        ),
-        'filename': 'Rastro.wsdl'
-    },
-    {
-        'url': (
-            'http://ws.correios.com.br/'
-            'calculador/CalcPrecoPrazo.asmx?WSDL'
-        ),
-        'filename': 'CalcPrecoPrazo.asmx'
-    },
-    {
-        'url': (
-            'https://webservice.correios.com.br/'
-            'service/rastro/Rastro_schema1.xsd'
-        ),
-        'filename': 'Rastro_schema1.xsd'
-    },
-]
+WSDL_DIR = os.path.join(os.path.dirname(__file__), 'wsdls')
 
 
-def create_file(filename, body, path):
+class WSDLDownloadError(Exception):
+    pass
+
+
+class WSDLUpdater:
+    def __init__(self, wsdl_dir=WSDL_DIR):
+        self.wsdl_dir = wsdl_dir
+
+
+def create_file(path, filename, body):
     file_path = os.path.join(path, filename)
-    logger.debug(
-        'Creating file: {filename}\n'
-        'Path: {path}'.format(filename=filename, path=path)
-    )
+    logger.debug('Creating file: {} Path: {}'.format(filename, path))
 
     with open(file_path, 'w+') as file:
         file.truncate()
         file.write(body)
 
-    logger.debug(
-        'Successfully create file: {filename}'.format(filename=filename)
-    )
+    logger.debug('Successfully create file: {}'.format(filename))
 
 
-def update_wsdl(path=MODULE_PATH):
+def update_wsdl(path=WSDL_DIR):
     for file_spec in SPECS:
-        logger.debug(
-            'Updating File: {filename}'.format(**file_spec)
-        )
+        url = file_spec['url']
+        filename = file_spec['filename']
 
-        response = requests.get(file_spec['url'])
+        logger.debug('Updating File: {} URL: {}'.format(filename, url))
 
+        response = requests.get(url)
         if response.status_code != 200:
-            logger.warning(
-                'Fail to access Correios: {url}'.format(**file_spec)
-            )
-            continue
+            raise WSDLDownloadError('Fail to download Correios: {}'.format(url))
 
-        create_file(file_spec['filename'], response.text, path=path)
+        create_file(path=path, filename=file_spec['filename'], body=response.text)
+
+
+def cli():
+    logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
+    update_wsdl()
+
+    parser = argparse.ArgumentParser(description='Updates Correios WSDL files')
+    parser.add_argument(
+        '-p',
+        '--path',
+        dest='wsdl_path',
+        default=WSDL_DIR,
+        help='Custom path where wsdl files will be saved.'
+    )
+    args = parser.parse_args()
+
+    print('Updating Correios WSDL')
+    print('Files will be saved on: {path}'.format(path=args.wsdl_path))
+    try:
+        update_wsdl(path=args.wsdl_path)
+        print('Updated with success')
+    except Exception as error:
+        print('Fail to update with error: {error}'.format(error=error))
 
 
 if __name__ == '__main__':
-    import sys
-
-    logging.basicConfig(stream=sys.stdout, level=logging.DEBUG)
-    update_wsdl()
+    cli()
