+++ correios/utils.py
@@ -14,12 +14,13 @@
 
 import os
 import re
+import pkg_resources
+
 from datetime import datetime
 from decimal import Decimal
 from itertools import chain
+from pathlib import Path
 from typing import Container, Iterable, Sized, Union
-
-import pkg_resources
 
 
 def capitalize_phrase(phrase: str) -> str:
@@ -99,10 +100,7 @@ def to_decimal(value: Union[Decimal, str, float], precision=2):
     return value.quantize(quantize)
 
 
-def get_wsdl_path(filename, path=None) -> str:
-    if not path:
-        resource_package = 'correios'
-        resource_path = '/'.join(('wsdls', filename))
-        return pkg_resources.resource_filename(resource_package, resource_path)
-
-    return os.path.join(path, filename)
+def get_resource_path(path) -> Path:
+    resource_package = 'correios'
+    resource_path = os.path.join("data", path)
+    return Path(pkg_resources.resource_filename(resource_package, resource_path))
