--- scripts/update-correios-wsdl
@@ -1,24 +1,0 @@
-#!/usr/bin/env python
-
-import argparse
-
-from correios.update_wsdl import update_wsdl, MODULE_PATH
-
-
-parser = argparse.ArgumentParser(description='Updates Correios WSDL files')
-parser.add_argument(
-    '-p',
-    '--path',
-    dest='wsdl_path',
-    default=MODULE_PATH,
-    help='Custom path where wsdl files will be saved.'
-)
-args = parser.parse_args()
-
-print('Updating Correios WSDL')
-print('Files will be saved on: {path}'.format(path=args.wsdl_path))
-try:
-    update_wsdl(path=args.wsdl_path)
-    print('Updated with success')
-except Exception as error:
-    print('Fail to update with error: {error}'.format(error=error))
