+++ tests/test_user_models.py
@@ -19,7 +19,6 @@ from datetime import datetime, timedelta, timezone
 import pytest
 from PIL.Image import Image
 
-from correios import DATADIR
 from correios.exceptions import (
     InvalidExtraServiceError,
     InvalidFederalTaxNumberError,
@@ -217,8 +216,8 @@ def test_basic_service():
     assert service.display_name == "SEDEX 10"
     assert service.description == "SEDEX 10"
     assert service.category == "SERVICO_COM_RESTRICAO"
-    assert service.get_symbol_filename() == os.path.join(DATADIR, "premium.gif")
-    assert service.get_symbol_filename("png") == os.path.join(DATADIR, "premium.png")
+    assert service.get_symbol_filename().endswith("/premium.gif")
+    assert service.get_symbol_filename("png").endswith("/premium.png")
     assert isinstance(service.symbol_image, Image)
     assert service.max_weight == 10000
